refactor: extract SplitSettings, CreateEditSplitProfile, ExportOptions, DateRangeSelector overlays from Dashboard
Extract 4 overlay JSX sections (~714 lines) from Dashboard.js into
dedicated components in components/Settings/:
- SplitSettingsOverlay.js (~242 lines, split profile selection/management)
- CreateEditSplitProfileOverlay.js (~347 lines, profile create/edit form)
- ExportOptionsOverlay.js (~166 lines, filtered/basic/full CSV export)
- DateRangeSelectorOverlay.js (~175 lines, preset and custom date ranges)

Dashboard.js: 7,449 -> 6,735 lines. Build passes.
